Restructure lobby events and add leave button
Lobby specific actions are now handled directly in the Lobby class.
The OnLobbyUpdated event can simply be defined per lobby in order to
delegate events to their respectful conections.

Create lobby logic has also been separated - Lobbies are created in the
LobbyService in order to manage the lobby state on a global level,
whereas players are created, removed and managed by the individual
lobbies since players are a lobby-specific concept.